requirements.txt
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
distro==1.9.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.14.0
groq==0.5.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jieba3k==0.35.1
joblib==1.4.0
lxml==5.2.1
lxml_html_clean==0.1.1
macholib==1.16.3
newspaper3k==0.2.8
nltk==3.8.1
openai==1.30.5
packaging==24.0
pillow==10.3.0
pydantic==2.7.0
pydantic_core==2.18.1
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
requests-file==2.0.0
setuptools==69.5.1
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
tinysegmenter==0.3
tldextract==5.1.2
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
